fix(safety): clamp PtySize, checked placeholder run, u64 image-compose offsets (cycle 760)
From the exhaustive audit — overflow/panic defense on runtime-reachable
paths (panic=abort makes these real crashes):
- Terminal::new opens the PTY via clamp_pty_dim (mirrors resize) so a wide/
  HiDPI grid can't wrap the u16 pixel-dimension multiply
- placeholder_cells uses 'if let' instead of expect() on runs.last_mut() so
  the PTY reader thread can't panic if the push logic is later refactored
- ImageData::compose computes byte offsets in u64 (cw/ch already clamp them
  in-bounds) to avoid u32 wraparound on very large frames

(Rejected the audit's 'critical' kitty.rs chunks_exact panic finding —
chunks_exact does not panic on non-multiple lengths; false positive.)

+++ crates/kettle-vt/src/image.rs
@@ -134,8 +134,13 @@ impl ImageData {
         let dst = std::sync::Arc::make_mut(&mut self.rgba);
         for row in 0..ch {
             for col in 0..cw {
-                let s = ((row * src.width + col) * 4) as usize;
-                let d = (((y + row) * self.width + x + col) * 4) as usize;
+                // Cycle 760: compute byte offsets in u64 so the multiply can't
+                // wrap u32 for very large frames. `cw`/`ch` already clamp the
+                // result in-bounds (x+col < width, y+row < height), so the
+                // cast back to usize is always valid.
+                let s = ((row as u64 * src.width as u64 + col as u64) * 4) as usize;
+                let d = (((y as u64 + row as u64) * self.width as u64 + x as u64 + col as u64) * 4)
+                    as usize;
                 if replace {
                     dst[d..d + 4].copy_from_slice(&src.rgba[s..s + 4]);
                     continue;
